+++ src/egmath.h
@@ -4,8 +4,8 @@
 /* public interface of egmath.c */
 
 void load15Flt3d3();
-void drawWorldObject(int16 shapeId, long worldX, long worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 scaleShift);
-void drawWorldLine(long worldX1, long worldY1, int alt1, long worldX2, long worldY2, int alt2, int color);
+void drawWorldObject(int16 shapeId, int32 worldX, int32 worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 scaleShift);
+void drawWorldLine(int32 worldX1, int32 worldY1, int16 alt1, int32 worldX2, int32 worldY2, int16 alt2, int color);
 void drawTargetView(int16 shapeId, int16 worldX, int16 worldY, int16 altitude, int16 objYaw, int16 objPitch, int16 objRoll, int16 mode, int16 shift);
 int16 shapeDataOffset(int16 shapeId);
 int16 clampRange(int16 value, int16 minVal, int16 maxVal);
